Added reverb g2 controller holder
How to properly use the controller mount. 

Due to the heavy rings on the g2 controller, I added a piece on top to wrap the wrist strap around for intense gameplay. The holder is snug enough to hold the controller safely however the notch on top of the ring is there just in case.