Seperate I2C function from haptics.ino
The function itself was moved into its own file for I2C to make it easier if someone else needs I2C support for other hardware in the future.

Currently it's still called in the haptics.ino with the PCA9685 setup code as that is the only thing to use it.